chore: reorganize a few imports to improve bundle size
