+++ apps/studio/components/interfaces/SQLEditor/SQLEditor.utils.ts
@@ -1,9 +1,7 @@
-// @ts-ignore
 import { NEW_SQL_SNIPPET_SKELETON, destructiveSqlRegex } from './SQLEditor.constants'
 import type { SqlSnippets, UserContent } from 'types'
 import { DiffType } from './SQLEditor.types'
 import { removeCommentsFromSql } from 'lib/helpers'
-import { stripIndent } from 'common-tags'
 
 export const createSqlSnippetSkeleton = ({
   id,
@@ -62,17 +60,17 @@ export function checkDestructiveQuery(sql: string) {
   return destructiveSqlRegex.some((regex) => regex.test(removeCommentsFromSql(sql)))
 }
 
-export const generateMigrationCliCommand = (id: string, name: string, isNpx = false) => stripIndent`
+export const generateMigrationCliCommand = (id: string, name: string, isNpx = false) => `
   ${isNpx ? 'npx ' : ''}supabase snippets download ${id} |
       ${isNpx ? 'npx ' : ''}supabase migration new ${name}
 `
 
-export const generateSeedCliCommand = (id: string, isNpx = false) => stripIndent`
+export const generateSeedCliCommand = (id: string, isNpx = false) => `
   ${isNpx ? 'npx ' : ''}supabase snippets download ${id} >> \\
       supabase/seed.sql
 `
 
-export const generateFileCliCommand = (id: string, name: string, isNpx = false) => stripIndent`
+export const generateFileCliCommand = (id: string, name: string, isNpx = false) => `
   ${isNpx ? 'npx ' : ''}supabase snippets download ${id} > \\
       ${name}.sql
 `
