+++ apps/studio/lib/helpers.ts
@@ -1,7 +1,5 @@
-import { post } from 'lib/common/fetch'
-import { API_URL, DEFAULT_MINIMUM_PASSWORD_STRENGTH, PASSWORD_STRENGTH } from 'lib/constants'
-import { toast } from 'react-hot-toast'
-import { v4 as _uuidV4 } from 'uuid'
+export { default as passwordStrength } from './password-strength'
+export { default as uuidv4 } from './uuid'
 
 export const tryParseJson = (jsonString: any) => {
   try {
@@ -48,10 +46,6 @@ export const removeJSONTrailingComma = (jsonString: string) => {
    * bracket ']' using a regular expression.
    */
   return jsonString.replace(/,\s*(?=[\}\]])/g, '')
-}
-
-export const uuidv4 = () => {
-  return _uuidV4()
 }
 
 export const timeout = (ms: number) => {
@@ -160,49 +154,6 @@ export const copyToClipboard = async (str: string | Promise<string>, callback = 
       .then(callback)
   } else {
     console.warn('Unable to copy to clipboard')
-  }
-}
-
-export async function passwordStrength(value: string) {
-  let message: string = ''
-  let warning: string = ''
-  let strength: number = 0
-
-  if (value && value !== '') {
-    if (value.length > 99) {
-      message = `${PASSWORD_STRENGTH[0]} Maximum length of password exceeded`
-      warning = `Password should be less than 100 characters`
-    } else {
-      // [Joshen] Unable to use RQ atm due to our Jest tests being in JS
-      const response = await post(`${API_URL}/profile/password-check`, { password: value })
-      if (!response.error) {
-        const { result } = response
-        const resultScore = result?.score ?? 0
-
-        const score = (PASSWORD_STRENGTH as any)[resultScore]
-        const suggestions = result.feedback?.suggestions
-          ? result.feedback.suggestions.join(' ')
-          : ''
-
-        message = `${score} ${suggestions}`
-        strength = resultScore
-
-        // warning message for anything below 4 strength :string
-        if (resultScore < DEFAULT_MINIMUM_PASSWORD_STRENGTH) {
-          warning = `${
-            result?.feedback?.warning ? result?.feedback?.warning + '.' : ''
-          } You need a stronger password.`
-        }
-      } else {
-        toast.error(`Failed to check password strength: ${response.error.message}`)
-      }
-    }
-  }
-
-  return {
-    message,
-    warning,
-    strength,
   }
 }
